Move towards a more general v1.0 interface (#79)
* restructure package to move circuits into a models.circuits module and model_io to save()/load() methods of the circuit class

* update examples to new package structure
